requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.
#
django-tagging-ext==0.2.0
django-tagging==0.3.1
django-extensions==0.4.1
django-oembed==0.1.1
django-piston==0.2.2

bleach
feedparser
